flask-sqlalchemy/requirements.txt
aniso8601==9.0.1
attrs==22.2.0
bcrypt==4.0.1
click==8.1.3
dnspython==2.2.1
email-validator==1.3.0
Flask==2.2.2
Flask-JWT-Extended==4.4.4
Flask-RESTful==0.3.9
flask-restx==1.0.3
Flask-SQLAlchemy==3.0.2
greenlet==2.0.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
MarkupSafe==2.1.1
marshmallow==3.19.0
packaging==22.0
PyJWT==2.6.0
pyrsistent==0.19.3
pytz==2022.7
six==1.16.0
SQLAlchemy==2.0.0rc1
typing_extensions==4.4.0
Werkzeug==2.2.2
